doc/en/user/requirements.txt
mkdocs-material>=9.5.9
mkdocs-include-markdown-plugin>=6.0.4
mkdocs-exclude>=1.0.2
mkdocs-macros-plugin>=1.0.4